git-svn-id: https://origsvn.digium.com/svn/asterisk/tags/1.4.26.2@216175 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ res/res_crypto.c
@@ -108,7 +108,7 @@ static int ssl_num_locks;
 
 static unsigned long ssl_threadid(void)
 {
-	return (unsigned long)pthread_self();
+	return pthread_self();
 }
 
 static void ssl_lock(int mode, int n, const char *file, int line)
